requirements.txt
pyTelegramBotAPI
pymupdf
python-docx
openai